split openapi into backend for frontend (bff) and external (ext) definitions
